% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glog}
\alias{glog}
\title{Generalized linear model constraint on hierarchical structure
by using overlapped group penalty}
\usage{
glog(y, x, g, v, lambda, hierarchy, family = "gaussian", rho = 10,
  scale = TRUE, eabs = 0.001, erel = 0.001, LL = 1, eta = 1.25,
  maxitr = 1000L)
}
\arguments{
\item{y}{response variable, in the format of matrix. When family is 
``gaussian'' or ``binomial'', \code{y} ought to
be a matrix of n by 1 for the observations; when family
is ``coxph'', y represents the survival objects, that is,
a matrix of n by 2, containing the survival time and the censoring status.
See \code{\link[survival]{Surv}}.}

\item{x}{a model matrix of dimensions n by p,in which the column represents 
the predictor variables.}

\item{g}{a numeric vector of group labels for the predictor variables.}

\item{v}{a numeric vector of binary values, represents whether or not the
predictor variables are penalized. Note that 1 indicates
penalization and 0 for not penalization.}

\item{lambda}{a numeric vector of three penalty parameters corresponding to L2 norm,
squared L2 norm, and L1 norm, respectively.}

\item{hierarchy}{a factor value in levels 0, 1, 2, which represent different
hierarchical structure within groups, respectively. 
When \code{hierarchy=0}, \eqn{\lambda_2} and \eqn{\lambda_3} are 
forced to be zeroes; when \code{hierarchy=1}, \eqn{\lambda_2} is 
forced to be zero; when \code{hierarchy=2}, there is no constraint 
on \eqn{\lambda}'s. See \code{\link{smog}}.}

\item{family}{a description of the distribution family for the response
variable variable. For continuous response variable,
family is ``gaussian''; for multinomial or binary response
variable, family is ``binomial''; for survival response
variable, family is ``coxph'', respectively.}

\item{rho}{the penalty parameter used in the alternating direction method
of multipliers algorithm (ADMM). Default is 10.}

\item{scale}{whether or not scale the design matrix. Default is \code{true}.}

\item{eabs}{the absolute tolerance used in the ADMM algorithm. Default is 1e-3.}

\item{erel}{the reletive tolerance used in the ADMM algorithm. Default is 1e-3.}

\item{LL}{initial value for the Lipschitz continuous constant for
approximation to the objective function in the Majorization-
Minimization (MM) (or iterative shrinkage-thresholding algorithm
(ISTA)). Default is 1.}

\item{eta}{gradient stepsize for the backtrack line search for the Lipschitz
continuous constant. Default is 1.25.}

\item{maxitr}{the maximum iterations for convergence in the ADMM algorithm.
Default is 500.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the model fitting. Default is \code{NULL}.}
}
\description{
Generalized linear model constraint on hierarchical structure
by using overlapped group penalty
}
\references{
\insertRef{ma2019structural}{smog}
}
\seealso{
\code{\link{cv.smog}}, \code{\link{smog.default}}, \code{\link{smog.formula}}, 
         \code{\link{predict.smog}}, \code{\link{plot.smog}}.
}
\author{
Chong Ma, \email{chong.ma@yale.edu}.
}
\keyword{internal}
