% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proxL1}
\alias{proxL1}
\title{proximal operator on L1 penalty}
\usage{
proxL1(x, lambda)
}
\arguments{
\item{x}{numeric value.}

\item{lambda}{numeric value for the L1 penalty parameter.}
}
\description{
proximal operator on L1 penalty
}
\keyword{internal}
