% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proxL2}
\alias{proxL2}
\title{proximal operator on L2 penalty}
\usage{
proxL2(x, lambda)
}
\arguments{
\item{x}{A numeric vector.}

\item{lambda}{numeric value for the L2 penalty parameter.}
}
\description{
proximal operator on L2 penalty
}
\keyword{internal}
