% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{smog}
\alias{smog}
\alias{print.smog}
\title{smog generic}
\usage{
smog(x, ...)

\method{print}{smog}(x, ...)
}
\arguments{
\item{x}{an object of class from ''smog''.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the cofficients table of the object x of class from ''smog''.
See \code{\link[base]{print}}.
}
\description{
smog generic
}
\references{
\insertRef{ma2019structural}{smog}
}
\seealso{
\code{\link{cv.smog}}, \code{\link{smog.default}}, \code{\link{smog.formula}},
\code{\link{predict.smog}}, \code{\link{plot.smog}}.
}
\keyword{internal}
