% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proxL1}
\alias{proxL1}
\title{L1 proximal operator}
\usage{
proxL1(x, lambda)
}
\arguments{
\item{x}{numeric value.}

\item{lambda}{numeric value for the L1 penalty parameter.}
}
\value{
A numeric value soft-thresholded by \eqn{\lambda},
which is \eqn{sign(x)(|x|-\lambda)_{+}}.
}
\description{
L1 proximal operator
}
\examples{
proxL1(2.0,0.5)

}
\references{
\insertRef{ma2019structural}{smog}
}
\author{
Chong Ma, \email{chongma8903@gmail.com}.
}
